backend/requirements-dev.txt
quart>=0.19,<1.0
quart-cors>=0.7,<1.0
motor>=3.6,<4.0
pymongo>=4.0,<5.0
httpx>=0.27,<1.0
redis[hiredis]>=5.0,<6.0
python-dotenv>=1.0,<2.0
pytest
pytest-asyncio
mongomock